checker for migrations

+++ db/migrate/20971124081428_change_page_tree_structure.rb
@@ -1,6 +1,6 @@
 class ChangePageTreeStructure < ActiveRecord::Migration[5.1]
   def change
-    ActiveRecord::Base.connection.tables.exclude?('slickr_pages')
+    if ActiveRecord::Base.connection.tables.exclude?('slickr_pages')
       remove_column :slickr_pages, :ancestry
       add_column :slickr_pages, :parent_id, :integer
       add_index :slickr_pages, :parent_id
